update coins list during swap process
